Have now included a more lightweight version of DiCER
This version now only needs a T1w image, and a functional image (with the mean) with NO intensity correction applied to the functional image.

This takes advantage of FSL's fast to do the segmentation as well as using AFNI tools to do detrending and high-pass filtering if desired!

If you include a tab-seperated-file in the standard of fMRIprep's v.1.1.0 confounds file you will then get a nice report of the regressors as well as FD and DVARS.